\name{MeanShift-package}
\alias{MeanShift-package}
\alias{MeanShift}
\docType{package}
\title{
\packageTitle{MeanShift}
}
\description{
\packageDescription{MeanShift}
}
\details{

\packageDESCRIPTION{MeanShift}
%%\packageIndices{MeanShift}
%%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MeanShift}

Maintainer: \packageMaintainer{MeanShift}
}
%%\references{
%%~~ Literature or other references for background information ~~
%%}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
%%~~ documentation directory ~~
%%\keyword{ package }
\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
\code{\link{bmsClustering}}, \code{\link{msClustering}}
}
%%\examples{
%%
%%}
